+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.4.7 (2017-04-12)
 * Fixed "`Phoenix.HTML.Form.Formex.CustomField.SelectAssoc/3` is undefined" errors
+* Added `opts` argument in `Formex.Builder/4`
 
 ## v0.4.6 (2017-04-10)
 * Fixed collections where items has ID as UUID instead of integer. `embedded_schema`